feat: first part of typescript migration
