--- server/exceptions/api-error.js
@@ -1,22 +1,0 @@
-module.exports = class ApiError extends Error {
-  status;
-  errors;
-
-  constructor(status, message, errors = []) {
-      super(message);
-      this.status = status;
-      this.errors = errors;
-  }
-
-  static UnauthorizedError() {
-      return new ApiError(401, 'Unauthorized error')
-  }
-
-  static BadRequest(message, errors = []) {
-      return new ApiError(400, message, errors);
-  }
-
-  static NotAllowed() {
-      return new ApiError(401, 'Not Allowed')
-  }
-}
