--- server/middlewares/auth-middleware.js
@@ -1,26 +1,0 @@
-const {validateAccessToken} = require("../service/token")
-const ApiError = require("../exceptions/api-error")
-
-module.exports = function (req, res, next) {
-  try {
-    const authorizationHeader = req.headers.authorization;
-    if(!authorizationHeader) {
-      return next(ApiError.UnauthorizedError())
-    }
-
-    const accessToken = authorizationHeader.split(' ')[1];
-    if(!accessToken) {
-      return next(ApiError.UnauthorizedError())
-    }
-
-    const userData = validateAccessToken(accessToken);
-    if(!userData) {
-      return next(ApiError.UnauthorizedError());
-    }
-
-    req.user = userData;
-    next()
-  } catch (error) {
-    return next(ApiError.UnauthorizedError())
-  }
-}
