--- server/middlewares/error-middleware.js
@@ -1,10 +1,0 @@
-const ApiError = require('../exceptions/api-error');
-
-module.exports = function (err, req, res, next) {
-    console.log(err);
-    if (err instanceof ApiError) {
-        return res.status(err.status).json({message: err.message})
-    }
-    return res.status(500).json({message: 'Unexpected error'})
-
-};
