--- server/models/Board.js
@@ -1,20 +1,0 @@
-const { Schema, model } = require("mongoose");
-
-const BoardSchema = new Schema({
-  title: {
-    type: String,
-    required: true,
-  },
-  owner: {
-    type: Schema.Types.ObjectId,
-    ref: "User",
-  },
-  users: [
-    {
-      type: Schema.Types.ObjectId,
-      ref: "User",
-    },
-  ],
-});
-
-module.exports = Board = model("Board", BoardSchema);
