--- server/models/Card.js
@@ -1,15 +1,0 @@
-const { Schema, model } = require("mongoose");
-
-const CardSchema = new Schema({
-  title: {
-    type: String,
-    required: true,
-  },
-  board: {
-    type: Schema.Types.ObjectId,
-    ref: "Board",
-  },
-  position: Number
-});
-
-module.exports = Card = model("Card", CardSchema);
