--- server/models/Token.js
@@ -1,12 +1,0 @@
-const { Schema, model } = require("mongoose");
-
-const TokenSchema = new Schema({
-  user: {
-    type: Schema.Types.ObjectId, ref: "User",
-  },
-  refreshToken: {
-    type: String, required: true,
-  },
-})
-
-module.exports = Token = model("Token", TokenSchema);
