--- server/models/User.js
@@ -1,22 +1,0 @@
-const { Schema, model } = require("mongoose");
-
-const UserSchema = new Schema({
-  email: {
-    type: String,
-    unique: true,
-    required: true,
-  },
-  password: {
-    type: String,
-    required: true,
-  },
-  isActivated: {
-    type: Boolean,
-    default: false,
-  },
-  activationLink: {
-    type: String,
-  },
-})
-
-module.exports = User = model("User", UserSchema);
