+++ server/src/models/User.ts
@@ -1,0 +1,23 @@
+import { Schema, model } from "mongoose";
+import { IUser } from "../types/types"
+
+const UserSchema = new Schema({
+  email: {
+    type: String,
+    unique: true,
+    required: true,
+  },
+  password: {
+    type: String,
+    required: true,
+  },
+  isActivated: {
+    type: Boolean,
+    default: false,
+  },
+  activationLink: {
+    type: String,
+  },
+})
+
+module.exports = model<IUser>("User", UserSchema);
